+++ src/webComponents/form/PrettyButton.ts
@@ -10,6 +10,8 @@ class PrettyButton extends HTMLElement {
     const variant = this.getAttribute("variant") ?? "default";
     const button = this.shadowRoot.querySelector("button");
     button.classList.add(variant);
+    
+    button.accessKey = this.accessKey;
   }
 
   set callback(value: () => void) {